Create separate file for study screen
